Specify the MaxIdleTime to prevent a WebException
Per the Litle XML Reference Guide section 2.3.2.4 the Litle servers will
close persistent connections after 10 seconds. Set the MaxIdleTime
property on the web request service point to prevent GetResponse() from
throwing [WebException: The underlying connection was closed: A
connection that was expected to be kept alive was closed by the server.]
after the connection is forcibly closed by the server.