Use application scope settings
Replace the user scope settings with application scoped settings to
support ASP.NET Web Services. User scope settings are not supported in
ASP.NET and will throw [ConfigurationErrorsException: The current
configuration system does not support user-scoped settings] if the SDK
is used in a web service. This change is not backwards compatible.